requirements.txt
alembic==1.8.1
attrs==22.1.0
click==8.1.3
colorama==0.4.5
decorator==5.1.1
flasgger==0.9.5
Flask==2.2.2
greenlet==1.1.3.post0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
Mako==1.2.3
MarkupSafe==2.1.1
mistune==2.0.4
prometheus-client==0.15.0
psycopg2-binary==2.9.4
pyrsistent==0.18.1
PyYAML==6.0
six==1.16.0
validators==0.20.0
Werkzeug==2.2.2
zipp==3.9.0
